+++ Back-end/routes/WishlistRoutes.js
@@ -54,10 +54,19 @@ router.put('/:id', async (req, res) => {
 });
 
 // delete wishlist item 
-router.delete('/:id', async (req, res) => {
+router.delete('/id', async (req, res) => {
     try {
-        const id = req.params.id;
+        const id = parseInt(req.params.id, 10);
+        if (isNaN(id)) {
+            return res.status(400).json({ error: 'Invalid ID' });
+        }
+        console.log(id);
         const deletedItem = await Wishlist.destroy({ where: { id } });
+
+        if (!deletedItem) {
+            return res.status(404).json({ error: 'Item not found' });
+        }
+
         res.status(200).json(deletedItem);
     }
     catch(error) {
